lab4_server/requirements.txt
flask
numpy
transformers
python-dotenv